lagoon-indexer/requirements.txt
web3
sqlalchemy
psycopg2-binary
python-dotenv
watchdog
pandas